% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_date.R
\name{knit_date}
\alias{knit_date}
\title{Custom Knit function for RStudio}
\usage{
knit_date(inputFile, ...)
}
\description{
This function customizes knitr output options.
It sets the outpot folder to up level folder output, (./output).
It generates .html file with date attached.
}
